Almost completed code. However there seems to be a bug when switching motors direction, the arduino stops responding. As if there was not enough power. But it does not reset. Still investigating.

+++ src/mpu.h
@@ -8,7 +8,7 @@ public:
   void initialize();
 
   void readIMUData();
-  inline float getAccAngle() const { return accAngle; }
+  inline double getAccAngle() const { return accAngle; }
   inline float getGyroRate() const { return gyroRate; }
 
   bool calibrateYGyro();
